Move replace_text_in and its support code into Dom and make it assert invariants!
