+++ crates/wysiwyg/src/composer_model.rs
@@ -18,7 +18,6 @@ pub mod example_format;
 pub mod format;
 mod format_inline_code;
 pub mod hyperlinks;
-pub mod join_nodes;
 pub mod lists;
 pub mod menu_state;
 pub mod replace_text;
