+++ crates/wysiwyg/src/composer_model/delete_text.rs
@@ -93,40 +93,6 @@ where
         }
     }
 
-    /// Deletes the given [to_delete] nodes and then removes any given parent nodes that became
-    /// empty, recursively.
-    /// Returns a list of all the nodes that were deleted
-    pub(crate) fn delete_nodes(
-        &mut self,
-        mut to_delete: Vec<DomHandle>,
-    ) -> Vec<DomHandle> {
-        let mut deleted = Vec::new();
-
-        // Delete in reverse order to avoid invalidating handles
-        to_delete.reverse();
-
-        // We repeatedly delete to ensure anything that became empty because
-        // of deletions is itself deleted.
-        while !to_delete.is_empty() {
-            // Keep a list of things we will delete next time around the loop
-            let mut new_to_delete = Vec::new();
-
-            for handle in to_delete.into_iter().filter(not_root) {
-                let index_in_parent = handle.index_in_parent();
-                let parent = self.state.dom.parent_mut(&handle);
-                parent.remove_child(index_in_parent);
-                adjust_handles_for_delete(&mut new_to_delete, &handle);
-                deleted.push(handle);
-                if parent.children().is_empty() {
-                    new_to_delete.push(parent.handle());
-                }
-            }
-
-            to_delete = new_to_delete;
-        }
-        deleted
-    }
-
     /// Removes the node at [cur_handle] and then will recursively delete any empty parent nodes
     /// until we reach the [top_handle] node.
     pub(crate) fn remove_and_clean_up_empty_nodes_until(
@@ -205,145 +171,5 @@ where
             // Default length for characters
             1
         }
-    }
-}
-
-fn not_root(handle: &DomHandle) -> bool {
-    !handle.is_root()
-}
-
-fn starts_with(subject: &DomHandle, object: &DomHandle) -> bool {
-    // Can't start with something longer than you
-    if subject.raw().len() < object.raw().len() {
-        return false;
-    }
-
-    // If any path element doesn't match we don't start with this
-    for (s, o) in subject.raw().iter().zip(object.raw().iter()) {
-        if s != o {
-            return false;
-        }
-    }
-
-    // All elements match, so we do start with it
-    true
-}
-
-fn adjust_handles_for_delete(
-    handles: &mut Vec<DomHandle>,
-    deleted: &DomHandle,
-) {
-    let mut indices_in_handles_to_delete = Vec::new();
-    let mut handles_to_replace = Vec::new();
-
-    let parent = deleted.parent_handle();
-    for (i, handle) in handles.iter().enumerate() {
-        if starts_with(handle, deleted) {
-            // We are the deleted node (or a descendant of it)
-            indices_in_handles_to_delete.push(i);
-        } else if starts_with(handle, &parent) {
-            // We are a sibling of the deleted node (or a descendant of one)
-
-            // If we're after a deleted node, reduce our index
-            let mut child_index = handle.raw()[parent.raw().len()];
-            let deleted_index = *deleted.raw().last().unwrap();
-            if child_index > deleted_index {
-                child_index -= 1;
-            }
-
-            // Create a handle with the adjusted index (but missing anything
-            // after the delete node's length).
-            let mut new_handle = parent.child_handle(child_index);
-
-            // Add back the rest of our original handle, unadjusted
-            for h in &handle.raw()[deleted.raw().len()..] {
-                new_handle = new_handle.child_handle(*h);
-            }
-            handles_to_replace.push((i, new_handle));
-        }
-    }
-
-    for (i, new_handle) in handles_to_replace {
-        handles[i] = new_handle;
-    }
-
-    indices_in_handles_to_delete.reverse();
-    for i in indices_in_handles_to_delete {
-        handles.remove(i);
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use crate::dom::DomHandle;
-    use crate::tests::testutils_composer_model::{cm, tx};
-
-    use super::*;
-
-    #[test]
-    fn starts_with_works() {
-        let h0123 = DomHandle::from_raw(vec![0, 1, 2, 3]);
-        let h012 = DomHandle::from_raw(vec![0, 1, 2]);
-        let h123 = DomHandle::from_raw(vec![1, 2, 3]);
-        let h = DomHandle::from_raw(vec![]);
-
-        assert!(starts_with(&h0123, &h012));
-        assert!(!starts_with(&h012, &h0123));
-        assert!(starts_with(&h012, &h012));
-        assert!(starts_with(&h012, &h));
-        assert!(!starts_with(&h123, &h012));
-        assert!(!starts_with(&h012, &h123));
-    }
-
-    #[test]
-    fn can_adjust_handles_when_removing_nodes() {
-        let mut handles = vec![
-            DomHandle::from_raw(vec![1, 2, 3]), // Ignored because before
-            DomHandle::from_raw(vec![2, 3, 4, 5]), // Deleted because inside
-            DomHandle::from_raw(vec![3, 4, 5]), // Adjusted because after
-            DomHandle::from_raw(vec![3]),       // Adjusted because after
-        ];
-
-        let to_delete = DomHandle::from_raw(vec![2]);
-
-        adjust_handles_for_delete(&mut handles, &to_delete);
-
-        assert_eq!(*handles[0].raw(), vec![1, 2, 3]);
-        assert_eq!(*handles[1].raw(), vec![2, 4, 5]);
-        assert_eq!(*handles[2].raw(), vec![2]);
-        assert_eq!(handles.len(), 3);
-    }
-
-    #[test]
-    fn can_adjust_handles_when_removing_nested_nodes() {
-        let mut handles = vec![
-            DomHandle::from_raw(vec![0, 9, 1, 2, 3]),
-            DomHandle::from_raw(vec![0, 9, 2, 3, 4, 5]),
-            DomHandle::from_raw(vec![0, 9, 3, 4, 5]),
-            DomHandle::from_raw(vec![0, 9, 3]),
-        ];
-
-        let to_delete = DomHandle::from_raw(vec![0, 9, 2]);
-
-        adjust_handles_for_delete(&mut handles, &to_delete);
-
-        assert_eq!(*handles[0].raw(), vec![0, 9, 1, 2, 3]);
-        assert_eq!(*handles[1].raw(), vec![0, 9, 2, 4, 5]);
-        assert_eq!(*handles[2].raw(), vec![0, 9, 2]);
-        assert_eq!(handles.len(), 3);
-    }
-
-    #[test]
-    fn delete_nodes_refuses_to_delete_root() {
-        let mut model = cm("|");
-        model.delete_nodes(vec![model.state.dom.document_handle()]);
-        assert_eq!(tx(&model), "|")
-    }
-
-    #[test]
-    fn delete_nodes_refuses_recursively_to_delete_root() {
-        let mut model = cm("a|");
-        model.delete_nodes(vec![DomHandle::from_raw(vec![0])]);
-        assert_eq!(tx(&model), "|")
     }
 }
